backend/requirements.txt
praw==7.7.1
pandas==2.1.4
python-dotenv==1.0.0
networkx==3.2.1
